Move sysutil.cpp under Platforms/Host.
