+++ src/Platforms/Host/sysutil.cpp
